Done alot of things. Add customers and preferences; has days looping; made money generate; pretty much all core game mechanics.

+++ Lemon.cs
@@ -6,9 +6,9 @@ using System.Threading.Tasks;
 
 namespace lemonadestand
 {
-    public class Lemon:Item
+    public class Lemon : Item
     {
-        public int sourness;
+        public double sourness;
         public Lemon(double price, int spoil, int sour)
         {
             sourness = sour;
